refactor(modules): replacing hooks inside api
